feat: add description for VIP-269
