--- simulations/vip-267/abi/LiquidateAndRedeem.json
@@ -1,243 +1,0 @@
-[
-    {
-        "inputs":
-        [
-            {
-                "internalType": "address",
-                "name": "owner_",
-                "type": "address"
-            },
-            {
-                "internalType": "contract IVBNB",
-                "name": "vNative",
-                "type": "address"
-            }
-        ],
-        "stateMutability": "nonpayable",
-        "type": "constructor"
-    },
-    {
-        "inputs":
-        [],
-        "name": "ApproveFailed",
-        "type": "error"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "uint256",
-                "name": "errCode",
-                "type": "uint256"
-            }
-        ],
-        "name": "LiquidationFailed",
-        "type": "error"
-    },
-    {
-        "inputs":
-        [],
-        "name": "NativeTransferFailed",
-        "type": "error"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "uint256",
-                "name": "errCode",
-                "type": "uint256"
-            }
-        ],
-        "name": "RedeemFailed",
-        "type": "error"
-    },
-    {
-        "inputs":
-        [],
-        "name": "ZeroAddressNotAllowed",
-        "type": "error"
-    },
-    {
-        "anonymous": false,
-        "inputs":
-        [
-            {
-                "indexed": true,
-                "internalType": "address",
-                "name": "previousOwner",
-                "type": "address"
-            },
-            {
-                "indexed": true,
-                "internalType": "address",
-                "name": "newOwner",
-                "type": "address"
-            }
-        ],
-        "name": "OwnershipTransferStarted",
-        "type": "event"
-    },
-    {
-        "anonymous": false,
-        "inputs":
-        [
-            {
-                "indexed": true,
-                "internalType": "address",
-                "name": "previousOwner",
-                "type": "address"
-            },
-            {
-                "indexed": true,
-                "internalType": "address",
-                "name": "newOwner",
-                "type": "address"
-            }
-        ],
-        "name": "OwnershipTransferred",
-        "type": "event"
-    },
-    {
-        "inputs":
-        [],
-        "name": "V_NATIVE",
-        "outputs":
-        [
-            {
-                "internalType": "contract IVBNB",
-                "name": "",
-                "type": "address"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [],
-        "name": "acceptOwnership",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "address",
-                "name": "borrower",
-                "type": "address"
-            },
-            {
-                "internalType": "contract IVBep20",
-                "name": "vTokenBorrowed",
-                "type": "address"
-            },
-            {
-                "internalType": "address",
-                "name": "receiver",
-                "type": "address"
-            }
-        ],
-        "name": "liquidateAndRedeemNative",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [],
-        "name": "owner",
-        "outputs":
-        [
-            {
-                "internalType": "address",
-                "name": "",
-                "type": "address"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [],
-        "name": "pendingOwner",
-        "outputs":
-        [
-            {
-                "internalType": "address",
-                "name": "",
-                "type": "address"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [],
-        "name": "renounceOwnership",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "address",
-                "name": "receiver",
-                "type": "address"
-            }
-        ],
-        "name": "sweepNative",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "contract IERC20Upgradeable",
-                "name": "token",
-                "type": "address"
-            },
-            {
-                "internalType": "address",
-                "name": "receiver",
-                "type": "address"
-            }
-        ],
-        "name": "sweepTokens",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "inputs":
-        [
-            {
-                "internalType": "address",
-                "name": "newOwner",
-                "type": "address"
-            }
-        ],
-        "name": "transferOwnership",
-        "outputs":
-        [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    },
-    {
-        "stateMutability": "payable",
-        "type": "receive"
-    }
-]
